+++ crates/cli/src/commands/pin.rs
@@ -5,16 +5,22 @@ use starbase_styles::color;
 use tracing::{debug, info};
 
 #[derive(Args, Clone, Debug)]
-pub struct GlobalArgs {
+pub struct PinArgs {
     #[arg(required = true, help = "ID of tool")]
     id: Id,
 
     #[arg(required = true, help = "Version or alias of tool")]
     spec: UnresolvedVersionSpec,
+
+    #[arg(
+        long,
+        help = "Add to the global user config instead of local .prototools"
+    )]
+    global: bool,
 }
 
 #[system]
-pub async fn global(args: ArgsRef<GlobalArgs>) -> SystemResult {
+pub async fn pin(args: ArgsRef<PinArgs>) -> SystemResult {
     let mut tool = load_tool(&args.id).await?;
 
     tool.manifest.default_version = Some(args.spec.clone());